Use TaskProvider concept
