+++ .github/workflows/Android-CI.yml
@@ -5,11 +5,11 @@ on:
     types: [assigned, opened, synchronize, reopened]
 
 env:
-  GRADLE_OPTS: -Dorg.gradle.daemon=false -Dkotlin.incremental=false -Dkotlin.compiler.execution.strategy=in-process -Dorg.gradle.internal.launcher.welcomeMessageEnabled=false
+  GRADLE_OPTS: -Dorg.gradle.internal.launcher.welcomeMessageEnabled=false
 
 jobs:
   buildDemo:
-    name: Build Android Demo JDK ${{ matrix.java_version }}
+    name: Build Android Demo JDK ${{ matrix.distribution }} ${{ matrix.java_version }}
     runs-on: ubuntu-latest
     defaults:
       run:
@@ -24,7 +24,7 @@ jobs:
         uses: actions/checkout@v2.4.0
         with:
           fetch-depth: 0
-      - name: Install JDK ${{ matrix.java_version }}
+      - name: Install JDK ${{ matrix.distribution }} ${{ matrix.java_version }}
         uses: actions/setup-java@v2.4.0
         with:
           distribution: ${{ matrix.distribution }}
@@ -32,4 +32,4 @@ jobs:
       - name: Install Android SDK
         uses: malinskiy/action-android/install-sdk@release/0.1.2
       - name: Build project
-        run: ./gradlew assembleDebug
+        run: ./gradlew assembleDebug -S
