+++ build.gradle.kts
@@ -68,6 +68,8 @@ dependencies {
 
 tasks.test {
     useJUnitPlatform()
+    systemProperty("junit.jupiter.testinstance.lifecycle.default", "per_method")
+    systemProperty("junit.jupiter.execution.parallel.enabled", "true")
 }
 
 configurations.all {
