+++ src/main/kotlin/org/jsonschema2dataclass/js2p/Js2pExtension.kt
@@ -3,7 +3,6 @@ package org.jsonschema2dataclass.js2p
 import org.gradle.api.NamedDomainObjectContainer
 import org.gradle.api.file.DirectoryProperty
 import org.gradle.api.tasks.Optional
-import org.gradle.api.tasks.OutputDirectory
 import org.gradle.api.tasks.PathSensitive
 import org.gradle.api.tasks.PathSensitivity
 import javax.inject.Inject
@@ -15,7 +14,6 @@ internal abstract class Js2pExtension @Inject constructor(
     @get: Optional
     abstract val executions: NamedDomainObjectContainer<Js2pConfiguration>
 
-    @get:OutputDirectory
     @get: PathSensitive(PathSensitivity.RELATIVE)
     abstract val targetDirectoryPrefix: DirectoryProperty
 }
